refactor and add live edit py
